# gin-blog-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/category": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "添加或修改分类",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "添加或修改分类",
                "parameters": [
                    {
                        "description": "添加或修改分类",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.AddOrEditCategoryReq"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.Response-model_Category"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据 ID 数组删除留言",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "删除留言（批量）",
                "parameters": [
                    {
                        "description": "留言 ID 数组",
                        "name": "ids",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.Response-int"
                        }
                    }
                }
            }
        },
...
        "/message/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据条件查询留言列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message"
                ],
                "summary": "条件查询留言列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "昵称",
                        "name": "nickname",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "审核状态",
                        "name": "is_review",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "当前页数",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数",
                        "name": "page_num",
                        "in": "query"
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.Response-handler_PageResult-model_Message"
                        }
                    }
                }
            }
        },
        "/message/review": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据 ID 数组修改审核状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message"
                ],
                "summary": "修改留言审核（批量）",
                "parameters": [
                    {
                        "description": "修改审核状态",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateReviewReq"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.Response-int"
                        }
                    }
                }
            }
        },
        "/operation/log": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除操作日志",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OperationLog"
                ],
                "summary": "删除操作日志",
                "parameters": [
                    {
                        "description": "操作日志ID列表",
                        "name": "ids",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.Response-int"
                        }
                    }
                }
            }
        },
        "/operation/log/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据条件查询获取操作日志列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OperationLog"
                ],
                "summary": "获取操作日志列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page_num",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "关键字",
                        "name": "keyword",
                        "in": "query"
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.Response-array_model_OperationLog"
                        }
                    }
                }
            }
        },
...
        "/role/list": {
            "get": {
                "description": "获取角色列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "summary": "获取角色列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "关键字",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.Response-handler_PageResult-model_RoleVO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.AddOrEditCategoryReq": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "handler.PageResult-model_Message": {
            "type": "object",
            "properties": {
                "page_data": {
                    "description": "分页数据",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Message"
                    }
                },
                "page_num": {
                    "description": "每页条数",
                    "type": "integer"
                },
                "page_size": {
                    "description": "上次页数",
                    "type": "integer"
                },
                "total": {
                    "description": "总条数",
                    "type": "integer"
                }
            }
        },
        "handler.PageResult-model_RoleVO": {
            "type": "object",
            "properties": {
                "page_data": {
                    "description": "分页数据",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.RoleVO"
                    }
                },
                "page_num": {
                    "description": "每页条数",
                    "type": "integer"
                },
                "page_size": {
                    "description": "上次页数",
                    "type": "integer"
                },
                "total": {
                    "description": "总条数",
                    "type": "integer"
                }
            }
        },
...
        "handler.Response-array_model_OperationLog": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.OperationLog"
                    }
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        },
...
        "handler.Response-handler_PageResult-model_Message": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/handler.PageResult-model_Message"
                        }
                    ]
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        },
        "handler.Response-handler_PageResult-model_RoleVO": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/handler.PageResult-model_RoleVO"
                        }
                    ]
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        },
...
        "handler.Response-int": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据",
                    "type": "integer"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        },
...
        "handler.Response-model_Category": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Category"
                        }
                    ]
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        },
...
        "handler.UpdateReviewReq": {
            "type": "object",
            "required": [
                "is_review"
            ],
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "is_review": {
                    "type": "integer",
                    "maximum": 1,
                    "minimum": 0
                }
            }
        },
        "model.Article": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "img": {
                    "type": "string"
                },
                "is_delete": {
                    "type": "integer"
                },
                "is_top": {
                    "type": "integer"
                },
                "original_url": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "model.Category": {
            "type": "object",
            "properties": {
                "articles": {
                    "description": "重写外键",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Article"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.Message": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ip_address": {
                    "type": "string"
                },
                "ip_source": {
                    "type": "string"
                },
                "is_review": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "speed": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.OperationLog": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ip_address": {
                    "type": "string"
                },
                "ip_source": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "opt_desc": {
                    "type": "string"
                },
                "opt_method": {
                    "type": "string"
                },
                "opt_module": {
                    "type": "string"
                },
                "opt_type": {
                    "type": "string"
                },
                "opt_url": {
                    "type": "string"
                },
                "request_method": {
                    "type": "string"
                },
                "request_param": {
                    "type": "string"
                },
                "response_data": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "model.RoleVO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_disable": {
                    "type": "boolean"
                },
                "label": {
                    "type": "string"
                },
                "menu_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "name": {
                    "type": "string"
                },
                "resource_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
    }
}